requirements.txt
flask
streamlit
torch
yolov5
opencv-python
numpy
pillow
requests
wtforms 